+++ EliteDangerousCompanionAppService/CompanionAppService.cs
@@ -49,21 +49,23 @@ namespace EliteDangerousCompanionAppService
 
         private static string serverRoot = "https://companion.orerve.net";
 
-        private Credentials credentials;
+        private CompanionAppCredentials credentials;
 
-        public CompanionAppService(Credentials credentials)
+        public CompanionAppService(CompanionAppCredentials credentials)
         {
             this.credentials = credentials;
         }
 
         ///<summary>Log in.  Returns credentials, or throws an exception if it fails</summary>
-        public static Credentials Login(string username, string password)
+        public static CompanionAppCredentials Login(string username, string password)
         {
-            Credentials credentials = null;
+            CompanionAppCredentials credentials = null;
+
             string location = serverRoot + "/user/login";
             // Send the request.
             HttpWebRequest request = (HttpWebRequest)WebRequest.Create(location);
             request.AllowAutoRedirect = false;  // Don't redirect or we lose the cookies
+
             request.UserAgent = "Mozilla/5.0 (iPhone; CPU iPhone OS 7_1_2 like Mac OS X) AppleWebKit/537.51.2 (KHTML, like Gecko) Mobile/11D257";
             request.ContentType = "application/x-www-form-urlencoded";
             request.Method = "POST";
@@ -99,19 +101,19 @@ namespace EliteDangerousCompanionAppService
                 Match companionAppMatch = Regex.Match(cookieHeader, @"CompanionApp=([^;]+)");
                 if (companionAppMatch.Success)
                 {
-                    if (credentials == null) { credentials = new Credentials(); }
+                    if (credentials == null) { credentials = new CompanionAppCredentials(); }
                     credentials.appId = companionAppMatch.Groups[1].Value;
                 }
                 Match machineIdMatch = Regex.Match(cookieHeader, @"mid=([^;]+)");
                 if (machineIdMatch.Success)
                 {
-                    if (credentials == null) { credentials = new Credentials(); }
+                    if (credentials == null) { credentials = new CompanionAppCredentials(); }
                     credentials.machineId = machineIdMatch.Groups[1].Value;
                 }
                 Match machineTokenMatch = Regex.Match(cookieHeader, @"mtk=([^;]+)");
                 if (machineTokenMatch.Success)
                 {
-                    if (credentials == null) { credentials = new Credentials(); }
+                    if (credentials == null) { credentials = new CompanionAppCredentials(); }
                     credentials.machineToken = machineTokenMatch.Groups[1].Value;
                 }
             }
@@ -130,7 +132,7 @@ namespace EliteDangerousCompanionAppService
         }
 
         ///<summary>Confirm a login.  Returns credentials, or throws an exception if it fails</summary>
-        public static Credentials Confirm(Credentials credentials, string code)
+        public static CompanionAppCredentials Confirm(CompanionAppCredentials credentials, string code)
         {
             var cookieContainer = new CookieContainer();
             AddCompanionAppCookie(cookieContainer, credentials);
@@ -174,19 +176,19 @@ namespace EliteDangerousCompanionAppService
                 Match companionAppMatch = Regex.Match(cookieHeader, @"CompanionApp=([^;]+)");
                 if (companionAppMatch.Success)
                 {
-                    if (credentials == null) { credentials = new Credentials(); }
+                    if (credentials == null) { credentials = new CompanionAppCredentials(); }
                     credentials.appId = companionAppMatch.Groups[1].Value;
                 }
                 Match machineIdMatch = Regex.Match(cookieHeader, @"mid=([^;]+)");
                 if (machineIdMatch.Success)
                 {
-                    if (credentials == null) { credentials = new Credentials(); }
+                    if (credentials == null) { credentials = new CompanionAppCredentials(); }
                     credentials.machineId = machineIdMatch.Groups[1].Value;
                 }
                 Match machineTokenMatch = Regex.Match(cookieHeader, @"mtk=([^;]+)");
                 if (machineTokenMatch.Success)
                 {
-                    if (credentials == null) { credentials = new Credentials(); }
+                    if (credentials == null) { credentials = new CompanionAppCredentials(); }
                     credentials.machineToken = machineTokenMatch.Groups[1].Value;
                 }
             }
@@ -244,19 +246,19 @@ namespace EliteDangerousCompanionAppService
                 Match companionAppMatch = Regex.Match(cookieHeader, @"CompanionApp=([^;]+)");
                 if (companionAppMatch.Success)
                 {
-                    if (credentials == null) { credentials = new Credentials(); }
+                    if (credentials == null) { credentials = new CompanionAppCredentials(); }
                     credentials.appId = companionAppMatch.Groups[1].Value;
                 }
                 Match machineIdMatch = Regex.Match(cookieHeader, @"mid=([^;]+)");
                 if (machineIdMatch.Success)
                 {
-                    if (credentials == null) { credentials = new Credentials(); }
+                    if (credentials == null) { credentials = new CompanionAppCredentials(); }
                     credentials.machineId = machineIdMatch.Groups[1].Value;
                 }
                 Match machineTokenMatch = Regex.Match(cookieHeader, @"mtk=([^;]+)");
                 if (machineTokenMatch.Success)
                 {
-                    if (credentials == null) { credentials = new Credentials(); }
+                    if (credentials == null) { credentials = new CompanionAppCredentials(); }
                     credentials.machineToken = machineTokenMatch.Groups[1].Value;
                 }
             }
@@ -276,7 +278,7 @@ namespace EliteDangerousCompanionAppService
             }
         }
 
-        private static void AddCompanionAppCookie(CookieContainer cookies, Credentials credentials)
+        private static void AddCompanionAppCookie(CookieContainer cookies, CompanionAppCredentials credentials)
         {
             var appCookie = new Cookie();
             appCookie.Domain = "companion.orerve.net";
@@ -286,7 +288,7 @@ namespace EliteDangerousCompanionAppService
             cookies.Add(appCookie);
         }
 
-        private static void AddMachineIdCookie(CookieContainer cookies, Credentials credentials)
+        private static void AddMachineIdCookie(CookieContainer cookies, CompanionAppCredentials credentials)
         {
             var machineIdCookie = new Cookie();
             machineIdCookie.Domain = ".companion.orerve.net";
@@ -296,7 +298,7 @@ namespace EliteDangerousCompanionAppService
             cookies.Add(machineIdCookie);
         }
 
-        private static void AddMachineTokenCookie(CookieContainer cookies, Credentials credentials)
+        private static void AddMachineTokenCookie(CookieContainer cookies, CompanionAppCredentials credentials)
         {
             var machineTokenCookie = new Cookie();
             machineTokenCookie.Domain = ".companion.orerve.net";