+++ EliteDangerousNetLogMonitor/NetLogMonitor.cs
@@ -8,7 +8,7 @@ namespace EliteDangerousNetLogMonitor
     /// <summary>A log monitor for the Elite: Dangerous netlog</summary>
     public class NetLogMonitor : LogMonitor
     {
-        public NetLogMonitor(string directory, Action<dynamic> callback) : base(directory, @"netLog", (result) => HandleNetLogLine(result, callback))
+        public NetLogMonitor(NetLogConfiguration configuration, Action<dynamic> callback) : base(configuration.path, @"netLog", (result) => HandleNetLogLine(result, callback))
         {
         }
 
@@ -25,29 +25,5 @@ namespace EliteDangerousNetLogMonitor
                 callback(result);
             }
         }
-
-        public static string ObtainDefaultPath()
-        {
-            String dataDir = Environment.GetEnvironmentVariable("AppData") + "\\EDDI";
-            Directory.CreateDirectory(dataDir);
-            string filename = dataDir + "\\productpath";
-            try
-            {
-                return File.ReadAllText(filename);
-            }
-            catch
-            {
-                return null;
-            }
-        }
-
-        public static void WritePath(string path)
-        {
-            String dataDir = Environment.GetEnvironmentVariable("AppData") + "\\EDDI";
-            Directory.CreateDirectory(dataDir);
-            string filename = dataDir + "\\productpath";
-            File.WriteAllText(filename, path);
-        }
     }
-
 }