+++ configuration/MainWindow.xaml.cs
@@ -1,0 +1,244 @@
+﻿using EliteDangerousCompanionAppService;
+using EliteDangerousDataDefinitions;
+using EliteDangerousNetLogMonitor;
+using EliteDangerousStarMapService;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace configuration
+{
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// </summary>
+    public partial class MainWindow : Window
+    {
+        public MainWindow()
+        {
+            InitializeComponent();
+
+            // Configure the Companion App tab
+            CompanionAppCredentials companionAppCredentials = CompanionAppCredentials.FromFile();
+            // See if the credentials work
+            CompanionAppService companionAppService = new CompanionAppService(companionAppCredentials);
+            try
+            {
+                Commander commander = companionAppService.Profile();
+                setUpCompanionAppComplete("Your connection to the companion app is operational, Commander " + commander.Name);
+            }
+            catch (Exception ex)
+            {
+                // Fall back to stage 1
+                setUpCompanionAppStage1();
+            }
+
+            // Configure the NetLog tab
+            NetLogConfiguration netLogConfiguration = NetLogConfiguration.FromFile();
+            netLogPathTextBox.Text = netLogConfiguration.path;
+
+            // Configure the EDSM tab
+            StarMapConfiguration starMapConfiguration = StarMapConfiguration.FromFile();
+            edsmApiKeyTextBox.Text = starMapConfiguration.apiKey;
+            edsmCommanderNameTextBox.Text = starMapConfiguration.commanderName;
+        }
+
+        private void companionAppNextClicked(object sender, RoutedEventArgs e)
+        {
+            // See if the user is entering their email address and password
+            if (companionAppEmailText.Visibility == Visibility.Visible)
+            {
+                // Stage 1 of authentication - login
+                string email = companionAppEmailText.Text.Trim();
+                string password = companionAppPasswordText.Text.Trim();
+                try
+                {
+                    CompanionAppCredentials companionAppCredentials = CompanionAppService.Login(email, password);
+                    companionAppCredentials.ToFile();
+                    setUpCompanionAppStage2();
+                }
+                catch (EliteDangerousCompanionAppAuthenticationException ex)
+                {
+                    companionAppText.Text = ex.Message;
+                }
+                catch (EliteDangerousCompanionAppErrorException ex)
+                {
+                    companionAppText.Text = ex.Message;
+                }
+                catch (Exception ex)
+                {
+                    companionAppText.Text = "Unexpected problem\r\nPlease report this at http://github.com/CmdrMcDonald/EliteDangerousDataProvider/issues\r\n" + ex;
+                }
+            }
+            else if (companionAppCodeText.Visibility == Visibility.Visible)
+            {
+                // Stage 2 of authentication - confirmation
+                try
+                {
+                    CompanionAppCredentials companionAppCredentials = CompanionAppCredentials.FromFile();
+                    companionAppCredentials = CompanionAppService.Confirm(companionAppCredentials, companionAppCodeText.Text.Trim());
+                    companionAppCredentials.ToFile();
+                    // All done - see if it works
+                    CompanionAppService companionAppService = new CompanionAppService(companionAppCredentials);
+                    Commander commander = companionAppService.Profile();
+                    setUpCompanionAppComplete("Your connection to the companion app is operational, Commander " + commander.Name);
+                }
+                catch (EliteDangerousCompanionAppAuthenticationException ex)
+                {
+                    setUpCompanionAppStage1(ex.Message);
+                }
+                catch (EliteDangerousCompanionAppErrorException ex)
+                {
+                    setUpCompanionAppStage1(ex.Message);
+                }
+                catch (Exception ex)
+                {
+                    setUpCompanionAppStage1("Unexpected problem\r\nPlease report this at http://github.com/CmdrMcDonald/EliteDangerousDataProvider/issues\r\n" + ex);
+                }
+            }
+        }
+
+        private void setUpCompanionAppStage1(string message = null)
+        {
+            if (message == null)
+            {
+                companionAppText.Text = "You do not have a connection to the companion app at this time.  Please enter your Elite: Dangerous email address and password below";
+            }
+            else
+            {
+                companionAppText.Text = message;
+            }
+
+            companionAppEmailLabel.Visibility = Visibility.Visible;
+            companionAppEmailText.Visibility = Visibility.Visible;
+            companionAppPasswordLabel.Visibility = Visibility.Visible;
+            companionAppPasswordText.Visibility = Visibility.Visible;
+            companionAppCodeLabel.Visibility = Visibility.Hidden;
+            companionAppCodeText.Visibility = Visibility.Hidden;
+            companionAppNextButton.Visibility = Visibility.Visible;
+        }
+
+        private void setUpCompanionAppStage2(string message = null)
+        {
+            if (message == null)
+            {
+                companionAppText.Text = "Please enter the verification code that should have been sent to your email address";
+            }
+            else
+            {
+                companionAppText.Text = message;
+            }
+
+            companionAppEmailLabel.Visibility = Visibility.Hidden;
+            companionAppEmailText.Visibility = Visibility.Hidden;
+            companionAppPasswordLabel.Visibility = Visibility.Hidden;
+            companionAppPasswordText.Visibility = Visibility.Hidden;
+            companionAppCodeLabel.Visibility = Visibility.Visible;
+            companionAppCodeText.Visibility = Visibility.Visible;
+            companionAppNextButton.Visibility = Visibility.Visible;
+        }
+
+        private void setUpCompanionAppComplete(string message = null)
+        {
+            if (message == null)
+            {
+                companionAppText.Text = "Complete";
+            }
+            else
+            {
+                companionAppText.Text = message;
+            }
+
+            companionAppEmailLabel.Visibility = Visibility.Hidden;
+            companionAppEmailText.Visibility = Visibility.Hidden;
+            companionAppPasswordLabel.Visibility = Visibility.Hidden;
+            companionAppPasswordText.Visibility = Visibility.Hidden;
+            companionAppCodeLabel.Visibility = Visibility.Hidden;
+            companionAppCodeText.Visibility = Visibility.Hidden;
+            companionAppNextButton.Visibility = Visibility.Hidden;
+        }
+
+        // Handle changes to NetLog tab
+        private void netLogPathChanged(object sender, TextChangedEventArgs e)
+        {
+            updateNetLogConfiguration();
+        }
+
+        private void updateNetLogConfiguration()
+        {
+            NetLogConfiguration netLogConfiguration = new NetLogConfiguration();
+            if (!String.IsNullOrWhiteSpace(netLogPathTextBox.Text))
+            {
+                netLogConfiguration.path = netLogPathTextBox.Text.Trim();
+            }
+            netLogConfiguration.ToFile();
+        }
+
+        private void netLogObtainClicked(object sender, RoutedEventArgs e)
+        {
+            List<string> processPaths = new Finder().GetPathFromProcess();
+            if (processPaths.Count != 0)
+            {
+                netLogPathTextBox.Text = processPaths[0] + @"\Logs";
+                updateNetLogConfiguration();
+            }
+            else
+            {
+                netLogText.Text = @"Unfortuantely we were unable to locate your product directory.  Please type in the location of your 'elite-dangerous-64' directory.  Possible locations include:";
+                List<string> paths = new Finder().FindInstallationPaths();
+                if (paths.Count == 0)
+                {
+                    paths.Add(Finder.DefProductsPath + @"\elite-dangerous-64");
+                    paths.Add(Finder.DefLauncherPath + @"\elite-dangerous-64");
+                    paths.Add(@"C:\Program Files (x86)\Steam\\SteamApps\common\Elite Dangerous\Products\elite-dangerous-64 (for Steam)");
+                }
+                foreach (string path in paths)
+                {
+                    netLogText.Text += "\r\n\r\n" + path;
+                }
+                netLogText.Text += "\r\n\r\nWhichever directory you select should contain a Logs directory and inside that a number of 'debugLog' files.";
+            }
+        }
+
+        // Handle changes to EDSM tab
+        private void edsmCommanderNameChanged(object sender, TextChangedEventArgs e)
+        {
+            updateEdsmConfiguration();
+        }
+
+        private void edsmApiKeyChanged(object sender, TextChangedEventArgs e)
+        {
+            updateEdsmConfiguration();
+        }
+
+        private void updateEdsmConfiguration()
+        {
+            StarMapConfiguration edsmConfiguration = new StarMapConfiguration();
+            if (!String.IsNullOrWhiteSpace(edsmApiKeyTextBox.Text))
+            {
+                edsmConfiguration.apiKey = edsmApiKeyTextBox.Text.Trim();
+            }
+            if (!String.IsNullOrWhiteSpace(edsmCommanderNameTextBox.Text))
+            {
+                edsmConfiguration.commanderName = edsmCommanderNameTextBox.Text.Trim();
+            }
+            edsmConfiguration.ToFile();
+
+        }
+
+        private void doneClicked(object sender, RoutedEventArgs e)
+        {
+            Application.Current.Shutdown();
+        }
+    }
+}